Delete Tickets Try

+++ deliriumm/tickets/views.py
@@ -128,12 +128,12 @@ def delete_ticket_request(request, requestid):
         
         if request.POST.get('yes', False):
             tr = get_object_or_404(TicketRequest, pk=requestid)
-
+            if os.path.exists(settings.MEDIA_ROOT + str(tr.comprobante)):
+                    os.remove(settings.MEDIA_ROOT + str(tr.comprobante))
             for ticket in tr.ticket_set.all():
-                # if os.path.exists(settings.MEDIA_ROOT + "/tickets/" + ticket.hash + '.jpg'):
-                os.remove(settings.MEDIA_ROOT + "/tickets/" + ticket.hash +'.jpg')
-                # if os.path.exists(settings.MEDIA_ROOT + str(tr.comprobante)):
-                os.remove(settings.MEDIA_ROOT + str(tr.comprobante))
+                if os.path.exists(settings.MEDIA_ROOT + "/tickets/" + ticket.hash + '.jpg'):
+                    os.remove(settings.MEDIA_ROOT + "/tickets/" + ticket.hash +'.jpg')
+
 
             get_object_or_404(TicketRequest, pk=requestid).delete()
             return HttpResponseRedirect(reverse("tickets:index"))
